requirements.txt
anthropic>=0.40.0
sse-starlette>=1.8.0
py-clob-client>=0.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets>=12.0
httpx>=0.26.0
pandas>=2.1.0
numpy>=1.26.0
rich>=13.7.0
pydantic>=2.5.0
python-binance>=1.0.19
asyncio-throttle>=1.0.2
tenacity>=8.2.3
loguru>=0.7.2
prometheus-client>=0.19.0
uvicorn>=0.25.0
fastapi>=0.109.0
